django/requirements/prod.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
Pillow==8.0.1
psycopg2-binary==2.8.6
gunicorn==20.0.4